+++ cmd/cremexplorer/config/config.go
@@ -1,3 +1,3 @@
 package config
 
-const Version = "0.9"
+const Version = "0.10"